season3/ewmkkpe/carbi-master/requirements.txt
alembic==0.9.6
blist==1.3.6
boto3==1.4.7
botocore==1.7.48
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
contextlib2==0.5.5
docutils==0.14
Flask==0.12.2
futures==3.2.0
gevent==1.2.2
google-api-core==0.1.4
google-api-python-client==1.6.5
google-auth==1.3.0
google-cloud-bigquery==0.29.0
google-cloud-core==0.28.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
greenlet==0.4.12
httplib2==0.10.3
idna==2.5
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
Mako==1.0.7
MarkupSafe==1.0
oauth2client==4.1.2
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyMySQL==0.7.2
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
raven==6.2.1
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
six==1.11.0
SQLAlchemy==1.1.15
ujson==1.35
uritemplate==3.0.0
urllib3==1.21.1
websocket-client==0.44.0
Werkzeug==0.13
gipc==0.6.0
